Extracted first round of shared test fixtures to lionrock-grpc-tests
