+++ lionrock-inmemory-server/release/start.bash
@@ -1,0 +1,4 @@
+#! /bin/bash
+
+set -xe;
+java -jar app.jar
