+++ lionrock-inmemory-server/src/main/java/io/github/panghy/lionrock/inmemory/InmemoryServer.java
@@ -1,0 +1,13 @@
+package io.github.panghy.lionrock.inmemory;
+
+import io.github.panghy.lionrock.proto.TransactionalKeyValueStoreGrpc;
+import org.lognet.springboot.grpc.GRpcService;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.stereotype.Component;
+
+@SpringBootApplication
+@GRpcService
+@Component
+public class InmemoryServer extends TransactionalKeyValueStoreGrpc.TransactionalKeyValueStoreImplBase {
+
+}
