+++ settings.gradle
@@ -4,4 +4,6 @@ include 'lionrock-foundationdb-server'
 include 'lionrock-foundationdb-client'
 include 'lionrock-cli'
 include 'lionrock-fdb-record-layer-client'
+include 'lionrock-inmemory-server'
+include 'lionrock-grpc-tests'
